# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Coopuertos",
        "description": "API REST para el sistema Coopuertos - Gestión de conductores, vehículos y propietarios",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "Servidor API v1"
        }
    ],
    "paths": {
...
        "/api/v1/vehiculos": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Listar vehículos",
                "description": "Retorna una lista paginada de vehículos con filtros opcionales",
                "operationId": "d0d6d5331f6af0db55a68379afe0671a",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items por página",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de vehículos"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Crear vehículo",
                "description": "Crea un nuevo vehículo",
                "operationId": "827c4ea51c30d822ae1f024a2ca87008",
                "responses": {
                    "201": {
                        "description": "Vehículo creado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/vehiculos/{id}": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Mostrar vehículo",
                "description": "Retorna la información de un vehículo específico",
                "operationId": "dbb0af856867b5ba2bb9e93761b01a6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Información del vehículo"
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Actualizar vehículo",
                "description": "Actualiza la información de un vehículo",
                "operationId": "89b73b8702c61c09ede3b6ce416fa684",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo actualizado exitosamente"
                    },
                    "422": {
                        "description": "Error de validación"
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Eliminar vehículo",
                "description": "Elimina un vehículo del sistema",
                "operationId": "a29557df650c35d23c8cb6016545ab5e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del vehículo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vehículo eliminado exitosamente"
                    },
                    "404": {
                        "description": "Vehículo no encontrado"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/vehiculos/search": {
            "get": {
                "tags": [
                    "Vehículos"
                ],
                "summary": "Buscar vehículos",
                "description": "Busca vehículos por placa, marca o modelo",
                "operationId": "cb767a3ba37a62f83b5daeac099fd179",
                "parameters": [
                    {
                        "name": "q",
                        "in": "query",
                        "description": "Término de búsqueda",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "minLength": 2
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Resultados de búsqueda"
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación de Laravel Sanctum",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "0": []
        }
    },
...
}